blaze_bot/blaze_bot/requirements.txt
requests
scikit-learn
python-telegram-bot==13.15
python-dotenv
